Added function to control the Response Message.
Responses are not longer send directly. The Response can be manually sent through the new public methods "response" and "sendResponse".
A timer with a configurable timeout will ensure that a response is always sent back to the sender.
The "hl7" Event now contains an object, with a message, message id, hl7-Object, and the default response code.

For Usage and changed behavior see Details in example.js.

Additionally, the mllpServer now emits events telling its state. (like connected, waiting, etc.)